+++ docs/changelog.rst
@@ -25,6 +25,7 @@ Additions to the :doc:`core-object-methods`.
     * Added the :py:meth:`khoros.core.Khoros.Message.define_context_id` method.
     * Added the :py:meth:`khoros.core.Khoros.Message.define_context_url` method.
 * Added the :py:meth:`khoros.core.Khoros.User.get_registered_users_count` method.
+* Added the :py:meth:`khoros.core.Khoros.Community.sso_enabled` method.
 
 Primary Modules
 ---------------
@@ -43,6 +44,7 @@ Additions to the :doc:`primary modules <primary-modules>`.
     * Added the :py:func:`khoros.objects.messages.define_context_id` function.
     * Added the :py:func:`khoros.objects.messages.define_context_url` function.
 * Added the :py:func:`khoros.objects.users.get_registered_users_count` function.
+* Added the :py:func:`khoros.structures.communities.sso_enabled` function.
 
 Supporting Modules
 ------------------